requirements.txt
# run pip install -r requirements.txt from FOE_Backend whith your venv active to install requirements

flask
flask_cors
Psycopg2==2.8.3
Flask-SqlAlchemy==2.4.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
SQLAlchemy-serializer==1.3.4.2
Flask-JWT-Extended==3.24.1
gunicorn==20.0.4
passlib==1.7.4
itsdangerous==1.1.0